Don't use upx.
UPX wasn't be used due to misconfiguration of the PyInstaller line.
While this is simple to fix, the PyInstalled upx compressed libraries
don't work with Python 3.x and appveyor.  Since it wasn't giving any
benefit, don't use it.

If the issue with upx and appveyor could be resolved, this could reduce
executables by 2 or 3 Mb.